Added cache to fix tracking multiple IDs
When attempting to run UpdateCharacterStatsAsync() where the same characterID appears multiple times within the batch of swipes, and FirstOrDefault will always return null because the newly added CharacterStats hasn't been saved yet. This causes an InvalidOperationError when attempting to run _context.CharacterStats.Add(stats) because it is attempting to add/track a new CharacterStats that already has been added/tracked but not yet saved.